Fix parallelism issues
