--- Sorting/ExternalMergeSort.cs
@@ -1,181 +1,0 @@
-﻿using System.Text;
-
-namespace Sorting;
-
-public interface IExternalSort
-{
-    void Apply(string filePath);
-}
-
-public class ExternalMergeSort : IExternalSort
-{
-    private const int NumbersLimit = 32_000_000; // 128M
-
-    public void Apply(string filePath)
-    {
-        try
-        {
-            var numbersBatch = new List<long>(NumbersLimit);
-            var chunksNum = GenerateSortedChunks(filePath, numbersBatch);
-            MergeSortedChunks(chunksNum);
-            MoveChunkToFile(0, filePath);
-        }
-        catch (Exception)
-        {
-            CleanUpChunks();
-            throw;
-        }
-    }
-
-    private static int GenerateSortedChunks(string filePath, List<long> numbersBatch)
-    {
-        int chunksNum = 0;
-        long? number;
-        using var numbersReader = new NumbersReader(filePath);
-        while ((number = numbersReader.Read()) != null)
-        {
-            numbersBatch.Add(number.Value);
-            if (numbersBatch.Count == NumbersLimit)
-            {
-                WriteSortedChunk(numbersBatch, chunksNum);
-                ++chunksNum;
-            }
-        }
-
-        if (numbersBatch.Count > 0)
-        {
-            WriteSortedChunk(numbersBatch, chunksNum);
-            ++chunksNum;
-        }
-
-        return chunksNum;
-    }
-
-    private static void WriteSortedChunk(List<long> chunk, long chunksNum)
-    {
-        chunk.Sort();
-        WriteChunk(chunk, chunksNum);
-        chunk.Clear();
-    }
-
-    private static void WriteChunk(IList<long> chunk, long chunkNum)
-    {
-        using var sw = CreatePerformantStreamWriter(GetChunkFileName(chunkNum));
-        foreach (var number in chunk)
-        {
-            sw.WriteLine(number);
-        }
-    }
-
-    private static StreamWriter CreatePerformantStreamWriter(string filePath)
-    {
-        return new(filePath, false, Encoding.Default, ushort.MaxValue + 1);
-    }
-
-    private static string GetChunkFileName(long chunkNum) => $"chunk-{chunkNum}.txt";
-
-    private static void MergeSortedChunks(int chunksNum)
-    {
-        while (chunksNum > 1)
-        {
-            Parallel.ForEach(GetChunksRange(chunksNum), range => { MergeChunks(range.chunkA, range.chunkB, range.chunkA / 2); });
-
-            if (chunksNum % 2 == 1) // uneven chunks num
-            {
-                var lastChunk = chunksNum - 1;
-                MoveChunkToFile(lastChunk, GetChunkFileName(lastChunk / 2));
-            }
-
-            chunksNum /= 2;
-        }
-    }
-
-    private static IEnumerable<(int chunkA, int chunkB)> GetChunksRange(int chunksNum)
-    {
-        for (int i = 0; i < chunksNum - 1; i += 2)
-        {
-            yield return (i, i + 1);
-        }
-    }
-
-    private static void MergeChunks(int chunkA, int chunkB, int chunkTarget)
-    {
-        var tempFileName = $"chunk-{chunkTarget}-temp.txt";
-        using (var targetWriter = CreatePerformantStreamWriter(tempFileName))
-        using (var bReader = new NumbersReader(GetChunkFileName(chunkB)))
-        using (var aReader = new NumbersReader(GetChunkFileName(chunkA)))
-        {
-            WriteSorted(aReader, bReader, targetWriter);
-
-            WriteToEnd(aReader, targetWriter);
-            WriteToEnd(bReader, targetWriter);
-        }
-
-        File.Delete(GetChunkFileName(chunkA));
-        File.Delete(GetChunkFileName(chunkB));
-
-        File.Move(tempFileName, GetChunkFileName(chunkTarget));
-    }
-
-    private static void WriteSorted(NumbersReader aReader, NumbersReader bReader, StreamWriter targetWriter)
-    {
-        for (long? a = aReader.Read(), b = bReader.Read();
-             a != null && b != null;)
-        {
-            if (a < b)
-            {
-                targetWriter.WriteLine(a);
-                a = aReader.Read();
-            }
-            else
-            {
-                targetWriter.WriteLine(b);
-                b = bReader.Read();
-            }
-        }
-    }
-
-    private static void WriteToEnd(NumbersReader reader, StreamWriter targetWriter)
-    {
-        for (long? k = reader.Read(); k != null; k = reader.Read())
-        {
-            targetWriter.WriteLine(k);
-        }
-    }
-
-    private static void MoveChunkToFile(int chunkNum, string filePath)
-    {
-        File.Delete(filePath);
-        File.Move(GetChunkFileName(chunkNum), filePath);
-    }
-
-    private static void CleanUpChunks()
-    {
-        var chunkPattern = GetChunkFileName(123).Replace("123", "*");
-        foreach (var chunkFile in Directory.GetFiles(".", chunkPattern))
-        {
-            File.Delete(chunkFile);
-        }
-    }
-}
-
-public sealed class NumbersReader : IDisposable
-{
-    private readonly TextReader _reader;
-
-    public NumbersReader(string filePath)
-    {
-        _reader = new StreamReader(filePath);
-    }
-
-    public long? Read()
-    {
-        var line = _reader.ReadLine();
-        return line == null ? null : long.Parse(line);
-    }
-
-    public void Dispose()
-    {
-        _reader.Dispose();
-    }
-}
